+++ src/components.d.ts
@@ -16,6 +16,12 @@ export namespace Components {
          */
         "type": AlertType;
     }
+    interface ZAnchorNavigation {
+        /**
+          * If enabled, the text of each anchor will be hidden unless for the current one or the one the user is hovering.
+         */
+        "hideUnselected": boolean;
+    }
     interface ZAppHeader {
         /**
           * The opening state of the drawer.
@@ -2330,6 +2336,12 @@ declare global {
         prototype: HTMLZAlertElement;
         new (): HTMLZAlertElement;
     };
+    interface HTMLZAnchorNavigationElement extends Components.ZAnchorNavigation, HTMLStencilElement {
+    }
+    var HTMLZAnchorNavigationElement: {
+        prototype: HTMLZAnchorNavigationElement;
+        new (): HTMLZAnchorNavigationElement;
+    };
     interface HTMLZAppHeaderElement extends Components.ZAppHeader, HTMLStencilElement {
     }
     var HTMLZAppHeaderElement: {
@@ -2980,6 +2992,7 @@ declare global {
     };
     interface HTMLElementTagNameMap {
         "z-alert": HTMLZAlertElement;
+        "z-anchor-navigation": HTMLZAnchorNavigationElement;
         "z-app-header": HTMLZAppHeaderElement;
         "z-app-switcher": HTMLZAppSwitcherElement;
         "z-app-topbar": HTMLZAppTopbarElement;
@@ -3096,6 +3109,12 @@ declare namespace LocalJSX {
           * alert variant type
          */
         "type"?: AlertType;
+    }
+    interface ZAnchorNavigation {
+        /**
+          * If enabled, the text of each anchor will be hidden unless for the current one or the one the user is hovering.
+         */
+        "hideUnselected"?: boolean;
     }
     interface ZAppHeader {
         /**
@@ -5461,6 +5480,7 @@ declare namespace LocalJSX {
     }
     interface IntrinsicElements {
         "z-alert": ZAlert;
+        "z-anchor-navigation": ZAnchorNavigation;
         "z-app-header": ZAppHeader;
         "z-app-switcher": ZAppSwitcher;
         "z-app-topbar": ZAppTopbar;
@@ -5576,6 +5596,7 @@ declare module "@stencil/core" {
     export namespace JSX {
         interface IntrinsicElements {
             "z-alert": LocalJSX.ZAlert & JSXBase.HTMLAttributes<HTMLZAlertElement>;
+            "z-anchor-navigation": LocalJSX.ZAnchorNavigation & JSXBase.HTMLAttributes<HTMLZAnchorNavigationElement>;
             "z-app-header": LocalJSX.ZAppHeader & JSXBase.HTMLAttributes<HTMLZAppHeaderElement>;
             "z-app-switcher": LocalJSX.ZAppSwitcher & JSXBase.HTMLAttributes<HTMLZAppSwitcherElement>;
             "z-app-topbar": LocalJSX.ZAppTopbar & JSXBase.HTMLAttributes<HTMLZAppTopbarElement>;